Add socket to http request (#20)
The socket field of an http request contains useful properties about the underlying socket for the request. This change exposes the properties for localAddress, localPort, remoteAddress, and remotePort.

This allows code handling an http request to lookup these properties. One example use case is identifying which incoming address a request was handled by. Another is identifying the remote source of a request - maybe to track requests from specific sources.