Remove requirement to initialize renderer before window load (#23)
### Public-Facing Changes

Removed the requirement for the renderer to be initialized before the
window's `load` event fires.

### Description

Relates to FG-2506 / https://github.com/foxglove/studio/pull/5070
Resolves https://github.com/foxglove/electron-socket/issues/14

Previously, this package required the renderer side to be initialized
via `await require("@foxglove/electron-socket/renderer").Create()`
**before** the window's `load` event. This was a simple mechanism for
the preloader side to ensure the renderer would be present to receive
its initial message containing a transferred MessagePort.

This change introduces a more complex handshake in order to allow the
renderer and preload side to be initialized in any order. The new flow
is:

#### If preloader is initialized first:
1. Preloader sets up a message listener.
1. Preloader sends `preloaderReady` message. The message is dropped
because the renderer is not set up.
1. Renderer sets up a message listener.
1. Renderer sends a `rendererReady` message.
1. Preloader receives that message, creates a MessagePort, and sends it
with a `preloaderInitialized` message.

#### If renderer is initialized first:
1. Renderer sets up a message listener.
1. Renderer sends `rendererReady` message. The message is dropped
because the preloader is not set up.
1. Preloader sets up a message listener.
1. Preloader sends a `preloaderReady` message.
1. The renderer receives that message, and sends _another_
`rendererReady` message.
1. The preloader receives that message, creates a MessagePort, and sends
it with a `preloaderInitialized` message.

With this change I also added a small test project that can be used to
manually verify working behavior. Adding integration tests would be a
good future goal.